backend/requirements.txt
fastapi
uvicorn
sqlmodel
asyncpg
python-dotenv
pydantic-settings
pyjwt
passlib==1.7.4
bcrypt==3.2.0
python-jose[cryptography]